pyproject.toml
[project]
name = "jobb-soeking"
version = "0.1.0"
description = "Job application assistant — research, generate, track."
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "anthropic>=0.83.0",
    "beautifulsoup4>=4.14.3",
    "jinja2>=3.1.6",
    "markdown>=3.10.2",
    "playwright>=1.58.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.3.3",
    "sqlalchemy>=2.0.47",
    "typer>=0.23.2",
]

[project.scripts]
jobb = "cli.main:app"

[tool.setuptools.packages.find]
include = ["cli*", "core*", "models*", "schemas*", "services*"]
